backend/requirements.txt
flask
flask-cors
requests
pandas
numpy
python-dotenv
apscheduler
yaml
